requirements.txt
asgiref==3.3.1
Django==3.1.4
django-leaflet==0.27.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
